Move tests for public APIs into the integration level.
That way one won't accidentally access private APIs, while staying
consistent with a standard Rust crate layout.

Personally, I really want to keep modules small and the `src/` tree
free of any extras.
