+++ crates/gitbutler-hunk-dependency/src/path.rs
@@ -835,7 +835,6 @@ fn insert_hunk_ranges(
 }
 
 #[cfg(test)]
-
 mod tests {
     use super::*;
