+++ crates/gitbutler-hunk-dependency/tests/mod.rs
@@ -1,1 +1,3 @@
-pub mod path;
+mod hunk;
+mod input;
+mod path;
